web-backend/hvz_scraper/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
docutils==0.18.1
idna==3.3
jsmin==3.0.0
lockfile==0.12.2
python-daemon==2.3.0
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
